feat: 추출검색바 컴포넌트화
