+++ app/components/title/Title.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../public/styles/variable" as *;
+@use "../../../public/styles/variable" as *;
 
 .title_container {
   text-align: center;
